Run payload validation as part of dataset endpoints
